Make Routine repository and controllers work with userId parameter

+++ src/main/java/com/itlbv/routineplanner/repository/jdbc/JdbcRoutineRepositoryImpl.java
@@ -33,33 +33,34 @@ public class JdbcRoutineRepositoryImpl implements RoutineRepository {
     }
 
     @Override
-    public Routine save(Routine routine) {
+    public Routine save(Routine routine, int userId) {
         MapSqlParameterSource map = new MapSqlParameterSource()
                 .addValue("id", routine.getId())
+                .addValue("user_id", userId)
                 .addValue("name", routine.getName());
         if (routine.isNew()) {
             Number newId = simpleInsert.executeAndReturnKey(map);
             routine.setId(newId.intValue());
         } else if (namedParameterJdbcTemplate.update(
-                "UPDATE routines SET name=:name WHERE id=:id", map) == 0) {
+                "UPDATE routines SET name=:name WHERE id=:id AND user_id=:user_id", map) == 0) {
             return null;
         }
         return routine;
     }
 
     @Override
-    public boolean delete(int id) {
-        return (jdbcTemplate.update("DELETE FROM routines WHERE id=?", id)) > 0;
+    public boolean delete(int id, int userId) {
+        return (jdbcTemplate.update("DELETE FROM routines WHERE id=? AND user_id=?", id, userId)) > 0;
     }
 
     @Override
-    public Routine get(int id) {
-        List<Routine> routines = jdbcTemplate.query("SELECT * FROM routines WHERE id=?", ROW_MAPPER, id);
+    public Routine get(int id, int userId) {
+        List<Routine> routines = jdbcTemplate.query("SELECT * FROM routines WHERE id=? AND user_id=?", ROW_MAPPER, id, userId);
         return DataAccessUtils.singleResult(routines);
     }
 
     @Override
-    public List<Routine> getAll() {
-        return jdbcTemplate.query("SELECT * FROM routines", ROW_MAPPER);
+    public List<Routine> getAll(int userId) {
+        return jdbcTemplate.query("SELECT * FROM routines WHERE user_id=?", ROW_MAPPER, userId);
     }
 }
